Documentation: Architecture/code outlining
Full type up of what architecture I think will be best along with way
too much hack assembly to get an idea for what we need to generate in
the codegen step. Should be rather straight forward to handle the code
just want to do it in a nice scalable way. Plus, this is the most
proper attempt at lexing, parsing, codegen that I've made so want to
have a decent structure as I'm already in knee deep.