Codegen: Handle push segment instructions
Best way to do this for now that I came up with is to just check the
next enum in the enum vector. Then it is only a few if-else if
statements to get the address that we need to address. After that it
is standard set of instructions for push.