Main: Increment codegen continue label
Continue is a label used in codegen for if-else style statements but it
needed a way to specify a difference between each of the continues. So
this value is just tacked onto the end of a continue and that makes it
individual because the value chages with each kind of instruction.