Codegen: Pass only the last token instead of whole vector
Pass only tokens[2] to functions instead of the whole tokens vector.
All the functions only make use of the second indexed token so it is
easier to use if we only pass it the token. This also allows us to pass
whatever last value we want even if the instruction has a different
token[2] value. I could pass 0 in as the last token despite the
instruction having say 8 as its token[2].

All tests pass.