Main: Handle directories of vm files
Now whole directories can be passed in as the argument to the
VMTranslator. Each directory is searched for .vm files and all the ones
that are found are inserted into a vector. The vector is then iterated
over and each file is open, translated, and closed. All vm files insert
their output into a singular output file based on the name of the
directory.