Main: ALL FUNCTION TEST PASSING WAHOOO!
Pass the individual file name to generate_hack_asm instead of the
overall output file name. Static variables use the name of the file
they are from, not the name of the single output file. If we use the
single output file name, static variables get overwritten when they
shouldn't. It's a great day!!!!