Main+Lexer: Correct regression caused in trim_path commit
Was passing a single file name to the generate_hack_asm function which
meant static variables all used the same label which leads to static
variables being overwritten. Added an extra parameter to the trim_path
function to specify a "full trim" which removes the path and the file
ending leaving just the file name. All tests pass.